+++ lewp-html/src/document_ext.rs
@@ -4,4 +4,8 @@
 pub trait DocumentExt {
     /// Converts the given node to an HTML string.
     fn into_html(self) -> String;
+    /// Parses the given HTML string to RcDom.
+    fn from_string(s: String) -> Result<Self, std::io::Error>
+    where
+        Self: Sized;
 }