+++ lewp-html/src/node_list.rs
@@ -1,0 +1,4 @@
+use crate::Node;
+
+/// A list of nodes.
+pub type NodeList = Vec<Node>;