+++ lewp/tests/page.rs
@@ -1,7 +1,7 @@
 use {
     lewp::{
         config::{ModuleConfig, PageConfig},
-        html::{api::*, Node, Nodes},
+        html::{api::*, Node, NodeList},
         Charset,
         LanguageTag,
         LewpError,
@@ -15,7 +15,7 @@ use {
 
 struct HelloWorld {
     pub config: ModuleConfig,
-    head_tags: Nodes,
+    head_tags: NodeList,
     data: String,
 }
 
@@ -40,7 +40,7 @@ impl From<ModuleConfig> for HelloWorld {
 }
 
 impl Module for HelloWorld {
-    fn head_tags(&self) -> &Nodes {
+    fn head_tags(&self) -> &NodeList {
         &self.head_tags
     }