+++ lewp/tests/sub-module.rs
@@ -4,7 +4,7 @@ mod modules {
     use {
         lewp::{
             config::ModuleConfig,
-            html::{api::*, Node, NodeExt, Nodes},
+            html::{api::*, Node, NodeExt, NodeList},
             LewpError,
             Module,
             Modules,
@@ -15,7 +15,7 @@ mod modules {
 
     pub struct Header {
         config: ModuleConfig,
-        head_tags: Nodes,
+        head_tags: NodeList,
         children: Modules,
         data: String,
     }
@@ -24,7 +24,7 @@ mod modules {
         pub fn new() -> Self {
             let mut instance = Self {
                 config: ModuleConfig::new(),
-                head_tags: Nodes::new(),
+                head_tags: NodeList::new(),
                 children: Modules::new(),
                 data: String::from("hello-world"),
             };
@@ -38,7 +38,7 @@ mod modules {
     }
 
     impl Module for Header {
-        fn head_tags(&self) -> &Nodes {
+        fn head_tags(&self) -> &NodeList {
             &self.head_tags
         }
 
@@ -79,7 +79,7 @@ mod modules {
 
     pub struct RandomHeadline {
         config: ModuleConfig,
-        head_tags: Nodes,
+        head_tags: NodeList,
         data: String,
     }
 
@@ -87,14 +87,14 @@ mod modules {
         pub fn new() -> Self {
             Self {
                 config: ModuleConfig::new(),
-                head_tags: Nodes::new(),
+                head_tags: NodeList::new(),
                 data: String::from("Wow this is dynamic!"),
             }
         }
     }
 
     impl Module for RandomHeadline {
-        fn head_tags(&self) -> &Nodes {
+        fn head_tags(&self) -> &NodeList {
             &self.head_tags
         }